pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = ""
authors = ["Martin Guindon <14781590+MartinGuindon@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
python-dotenv = "^0.21.0"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

